requirements.txt
netmiko==3.4.0
paramiko==2.7.2
pycryptodome==3.11.0
pandas==1.3.3